refactor: declara enum dentro da interface IMarvelContent
